% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api_domain}
\alias{api_domain}
\title{Function to get API access token}
\usage{
api_domain()
}
\value{
Auth0 API Endpoint domain \code{Sys.getenv("AUTH0_DOMAIN")}
}
\description{
This function returns your auth0 account endpoint.
}
