% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api_token}
\alias{api_token}
\title{Function to get API access token}
\usage{
api_token()
}
\value{
Auth0 API token from \code{Sys.getenv("AUTH0_API_KEY")}
}
\description{
To get the application token see: `https://auth0.com/docs/api/management/v2`
}
