% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_token}
\alias{generate_token}
\title{Generate Bearer auth token}
\usage{
generate_token(client_id, client_secret, .domain = api_domain())
}
\arguments{
\item{client_id}{The OAuth application id}

\item{client_secret}{The OAuth application secret}

\item{.domain}{By default \code{\link{api_domain}}}
}
\value{
Auth0 Bearer token
}
\description{
This function allows to generate new auth token base on client id and secret
If you looking for more infromation see: \url{https://auth0.com/docs/api/management/v2/tokens}
}
