% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_user}
\alias{get_user}
\title{Get a user endpoint}
\usage{
get_user(user_id, fields = NULL, include_fields = NULL, ...)
}
\arguments{
\item{user_id}{The user_id of the user to retrieve}

\item{fields}{A comma separated list of fields to include or exclude (depending on include_fields) from the result, empty to retrieve all fields}

\item{include_fields}{true if the fields specified are to be included in the result, false otherwise. Defaults to true}

\item{...}{all others settings \link{auth0} settings.}
}
\value{
Answer from the Auth0 API as \code{aut0_api} object.
}
\description{
\url{https://auth0.com/docs/api/management/v2#!/Users/get_users_by_id}
 This endpoint can be used to retrieve user details given the \code{user_id}.
}
