% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_domain}
\alias{set_domain}
\title{Function to set AUTH0_DOMAIN env variable}
\usage{
set_domain(domain)
}
\description{
Function to set AUTH0_DOMAIN env variable
}
\examples{
\dontrun{
  set_domain("https://auth0-org-domain.aut0.com")
}
}
