% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_token}
\alias{set_token}
\title{Function to set AUTH0_API_KEY env variable}
\usage{
set_token(api_token)
}
\arguments{
\item{api_token}{The Bearer API Auth0 token}
}
\description{
Function to set AUTH0_API_KEY env variable
}
