% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_user}
\alias{get_user}
\title{Get a user endpoint}
\usage{
get_user(user_id, fields = NULL, include_fields = NULL, ...)
}
\arguments{
\item{user_id}{The id of the user to retrieve}

\item{fields}{A comma separated list of fields to include or exclude (depending on include_fields) from the result.
If left empty then all fields are retrieved.}

\item{include_fields}{"true" if the fields specified are to be included in the result, "false" otherwise. Default is "true".}

\item{...}{other \link{auth0} settings.}
}
\value{
Response from the Auth0 API as \code{aut0_api} object.
}
\description{
The endpoint: \url{https://auth0.com/docs/api/management/v2#!/Users/get_users_by_id} can be used to retrieve user details given the \code{user_id}.
}
\examples{
\dontrun{
  get_user("google-oauth2|123123123213123123123")
}
}
